+++ API/MigrationGeneratorInterface.php
@@ -10,10 +10,11 @@ interface MigrationGeneratorInterface
     /**
      * Generate a migration
      *
-     * @param string $identifier
+     * @param string $matchType
+     * @param string|string[] $matchValue
      * @param string $mode
      * @return array Migration data
      * @throws \Exception
      */
-    public function generateMigration($identifier, $mode);
+    public function generateMigration($matchType, $matchValue, $mode);
 }